Add dataset implementations from codrep and bug2fix
This commit adds 2 things. First, implementations for the codrep and
bug2fix datasets. I tried using a similar interface to the datasets
found in torchvision and pytorch in general.
Second, I refactored the code for generating the buggy context, so that
it works better with JPype in python. JPype doesn't manage System.exit()
very well, and also it's super easy to send String(s) to the JVM using
JPype, so I changed the signatures of the relevant methods in the java
code.